refactor: nest config in guest command
this is to allow for future guest commands
