+++ cli/command/commands/commands.go
@@ -9,6 +9,7 @@ import (
 	_ "github.com/Telmate/proxmox-api-go/cli/command/delete"
 	_ "github.com/Telmate/proxmox-api-go/cli/command/example"
 	_ "github.com/Telmate/proxmox-api-go/cli/command/get"
+	_ "github.com/Telmate/proxmox-api-go/cli/command/get/guest"
 	_ "github.com/Telmate/proxmox-api-go/cli/command/get/id"
 	_ "github.com/Telmate/proxmox-api-go/cli/command/guest"
 	_ "github.com/Telmate/proxmox-api-go/cli/command/guest/qemu"
